Intensity Map Stretching
Introduce intensity map stretching into the intensity analysis section.
Because sometimes maxima get softened out when we use cv::Scharr to
calculate intensity curve derivatives via the gaussian smoothing it
performs, we need to stretch out the middle 7 pixels of the curve to
something larger.

However, there isn't a lot of data in these 7 pixels (approx 7 voxels
worth of data), so this doesn't seem to be a good long-term solution.

When converting stretched image coordinates back to voxel units, a
two-phase lookup process has to be performed:

  1) Convert stretched image coordinates back to reslice coordinates

  2) Convert reslice coordinates to voxel coordinates

The first transformation is done in the same way that the
resliceToVoxelCoord function in Reslice does it.